requirements_testing.txt
pytest==8.3.3
pytest-cov==5.0.0
moto[all]==5.0.13
ruff==0.5.7
mypy==1.11.2
types-requests
boto3-stubs[s3]

